Pull correct imported target from MBH cmake package.
This change depends on the MBH change to export the include directory
in order for the build to work; however, it fixes an underlying bug in
the invocation of target_link_libraries.

MBH is written such that it exports cmake variables under the prefix "MBH::"
Before this change I believe cmake was finding MBH via some system
wide configuration or the commented out lines removed in this change.

Paired with the change to MBH this "just works" and cmake finds MBH's
include and lib directories through MBH's exported .cmake file in the
way that is normal for cmake.